moved some basic utils out of fitgrains to coreutil
