+++ hexrd/__init__.py
@@ -72,7 +72,8 @@ if USE_NUMBA:
 del _readenv
 
 
-doc_url = 'latest' if 'dev' in __version__ else 'v%s' % __version__
+#doc_url = 'latest' if 'dev' in __version__ else 'v%s' % __version__
+doc_url = 'latest'
 doc_url = 'http://hexrd.readthedocs.org/en/%s' % doc_url
