+++ hexrd/xrd/xrdutil.py
@@ -72,7 +72,9 @@ from hexrd.xrd import transforms as xf
 from hexrd.xrd import transforms_CAPI as xfcapi
 
 from hexrd.xrd import distortion
-#from hexrd.fitgrains import get_frames, get_instrument_parameters
+
+#from hexrd.cacheframes import get_frames
+#from hexrd.coreutil import get_instrument_parameters
 
 try:
     from femODFUtil import pfig as pfigPkg
@@ -1726,11 +1728,11 @@ class EtaOmeMaps(object):
 # not ready # class BaseEtaOme(object):
 # not ready #     """
 # not ready #     eta-ome map base class derived from new YAML config
-# not ready #
+# not ready # 
 # not ready #     ...for now...
-# not ready #
+# not ready # 
 # not ready #     must provide:
-# not ready #
+# not ready # 
 # not ready #     self.dataStore
 # not ready #     self.planeData
 # not ready #     self.iHKLList
@@ -1738,7 +1740,7 @@ class EtaOmeMaps(object):
 # not ready #     self.omeEdges # IN RADIANS
 # not ready #     self.etas     # IN RADIANS
 # not ready #     self.omegas   # IN RADIANS
-# not ready #
+# not ready # 
 # not ready #     This wrapper will provide all but dataStore.
 # not ready #     """
 # not ready #     def __init__(self, cfg, reader=None, eta_step=None):
@@ -1748,34 +1750,34 @@ class EtaOmeMaps(object):
 # not ready #         """
 # not ready #         self.cfg = cfg
 # not ready #         self.instr_cfg = get_instrument_parameters(cfg)
-# not ready #
+# not ready # 
 # not ready #         # currently hard-coded to do reader from npz frame cache
 # not ready #         # kwarg *MUST* be 'new' style reader
 # not ready #         if reader is None:
 # not ready #             self.__reader = get_frames(reader, self.cfg)
 # not ready #         else:
 # not ready #             self.__reader = reader
-# not ready #
+# not ready # 
 # not ready #         # set eta_step IN DEGREES
 # not ready #         if eta_step is None:
 # not ready #             self._eta_step = self.cfg.image_series.omega.step
 # not ready #         else:
 # not ready #             self._eta_step = abs(eta_step) # just in case negative...
-# not ready #
+# not ready # 
 # not ready #         material_list = cPickle.load(open(cfg.material.definitions, 'r'))
 # not ready #         material_names = [material_list[i].name for i in range(len(material_list))]
 # not ready #         material_dict = dict(zip(material_names, material_list))
 # not ready #         self.planeData = material_dict[cfg.material.active].planeData
-# not ready #
+# not ready # 
 # not ready #         self._iHKLList = None
-# not ready #
+# not ready # 
 # not ready #         self._etaEdges = None
 # not ready #         self._omeEdges = None
 # not ready #         self._etas = None
 # not ready #         self._omegas = None
-# not ready #
+# not ready # 
 # not ready #         return
-# not ready #
+# not ready # 
 # not ready #     @property
 # not ready #     def iHKLList(self):
 # not ready #         return self._iHKLList
@@ -1786,7 +1788,7 @@ class EtaOmeMaps(object):
 # not ready #         """
 # not ready #         if ids is not None:
 # not ready #             assert hasattr(ids, '__len__'), "ids must be a list or list-like object"
-# not ready #
+# not ready # 
 # not ready #         # start with all available
 # not ready #         active_hkls = range(pd.hkls.shape[1])
 # not ready #         # check cfg file
@@ -1795,7 +1797,7 @@ class EtaOmeMaps(object):
 # not ready #         active_hkls = active_hkls if temp == 'all' else temp
 # not ready #         # override with hkls from command line, if specified
 # not ready #         return ids if ids is not None else active_hkls
-# not ready #
+# not ready # 
 # not ready #     @property
 # not ready #     def omegas(self):
 # not ready #         return self._omegas
@@ -1808,11 +1810,11 @@ class EtaOmeMaps(object):
 # not ready #         ome_start = self.__reader[1][0]
 # not ready #         ome_step  = self.__reader[1][1]
 # not ready #         return ome_step*(num.arange(num_ome) + 0.5) + ome_start
-# not ready #
+# not ready # 
 # not ready #     @property
 # not ready #     def eta_step(self):
 # not ready #         return self._eta_step
-# not ready #
+# not ready # 
 # not ready #     @property
 # not ready #     def etas(self):
 # not ready #         return self._etas
@@ -1820,12 +1822,12 @@ class EtaOmeMaps(object):
 # not ready #     def etas(self):
 # not ready #         """
 # not ready #         range is forced to be [-180, 180] for now, so step must be positive
-# not ready #
+# not ready # 
 # not ready #         step is same as omega unless specified (in degrees)
 # not ready #         """
 # not ready #         num_eta = int(360/float(abs(self.eta_step)))
 # not ready #         return num.radians(self.eta_step)*(num.arange(num_eta) + 0.5) - num.pi
-# not ready #
+# not ready # 
 # not ready #     @property
 # not ready #     def omeEdges(self):
 # not ready #         return self._omeEdges
@@ -1833,28 +1835,28 @@ class EtaOmeMaps(object):
 # not ready #     def omeEdges(self):
 # not ready #         ome_step = self.omegas[1] - self.omegas[0] # same as self.__reader[1][1]
 # not ready #         return num.hstack([self.omegas - 0.5*ome_step, self.omegas[-1] + 0.5*ome_step])
-# not ready #
+# not ready # 
 # not ready #     @property
 # not ready #     def etaEdges(self):
 # not ready #         return self._etaEdges
 # not ready #     @etaEdges.getter
 # not ready #     def etaEdges(self):
 # not ready #         return num.hstack([self.etas - 0.5*eta_step, self.etas[-1] + 0.5*eta_step])
-# not ready #
+# not ready # 
 # not ready # class EtaOmeMaps(BaseEtaOme):
 # not ready #     """
 # not ready #     """
 # not ready #     def __init__(self, cfg, reader=None, eta_step=None,
 # not ready #                  omega=0., tVec_s=num.zeros(3),
 # not ready #                  npdiv=2):
-# not ready #
+# not ready # 
 # not ready #         # first init the base class
 # not ready #         super( EtaOmeMaps, self ).__init__(cfg, reader=reader, eta_step=eta_step)
-# not ready #
+# not ready # 
 # not ready #         # grac relevant tolerances for patches
 # not ready #         tth_tol = num.degrees(self.planeData.tThWidth)
 # not ready #         eta_tol = num.degrees(abs(self.etas[1]-self.etas[0]))
-# not ready #
+# not ready # 
 # not ready #         # grab distortion
 # not ready #         if instr_cfg['detector']['distortion']['function_name'] is None:
 # not ready #             distortion = None
@@ -1872,28 +1874,28 @@ class EtaOmeMaps(object):
 # not ready #             ])
 # not ready #         pixel_pitch = instr_cfg['detector']['pixels']['size']
 # not ready #         chi = self.instr_cfg['oscillation_stage']['chi'] # in DEGREES
-# not ready #
+# not ready # 
 # not ready #         # 6 detector affine xform parameters
 # not ready #         rMat_d = makeDetectorRotMat(detector_params[:3])
 # not ready #         tVec_d = detector_params[3:6]
-# not ready #
+# not ready # 
 # not ready #         # 'dummy' sample frame rot mat
 # not ready #         rMats_s = makeOscillRotMat(num.radians([chi, omega]))
-# not ready #
+# not ready # 
 # not ready #         # since making maps for all eta, must hand trivial crystal params
 # not ready #         rMat_c = np.eye(3)
 # not ready #         tVec_c = np.zeros(3)
-# not ready #
+# not ready # 
 # not ready #         # make angle arrays for patches
 # not ready #         neta = len(self.etas)
 # not ready #         nome = len(reader[0])
-# not ready #
+# not ready # 
 # not ready #         # make full angs list
 # not ready #         angs = [num.vstack([tth*num.ones(neta),
 # not ready #                            etas,
 # not ready #                            num.zeros(nome)])
 # not ready #                 for tth in self.planeData.getTTh()]
-# not ready #
+# not ready # 
 # not ready #         """SET MAPS CONTAINER AS ATTRIBUTE"""
 # not ready #         self.dataStore = num.zeros((len(angs), nome, neta))
 # not ready #         for i_ring in range(len(angs)):
@@ -1902,7 +1904,7 @@ class EtaOmeMaps(object):
 # not ready #             xydet_ring = xfcapi.gvecToDetectorXY(gVec_ring_l,
 # not ready #                                                  rMat_d, rMat_s, rMat_c,
 # not ready #                                                  tVec_d, tVec_s, tVec_c)
-# not ready #
+# not ready # 
 # not ready #             if distortion is not None:
 # not ready #                 det_xy = distortion[0](xydet_ring,
 # not ready #                                        distortion[1],
@@ -1911,7 +1913,7 @@ class EtaOmeMaps(object):
 # not ready #                                       rMat_d, rMat_s,
 # not ready #                                       tVec_d, tVec_s, tVec_c,
 # not ready #                                       distortion=distortion)
-# not ready #
+# not ready # 
 # not ready #             patches = make_reflection_patches(self.instr_cfg,
 # not ready #                                               angs[i_ring].T[:, :2], ang_ps,
 # not ready #                                               omega=None,
@@ -1919,7 +1921,7 @@ class EtaOmeMaps(object):
 # not ready #                                               distortion=distortion,
 # not ready #                                               npdiv=npdiv, quiet=False,
 # not ready #                                               compute_areas_func=gutil.compute_areas)
-# not ready #
+# not ready # 
 # not ready #             for i in range(nome):
 # not ready #                 this_frame = num.array(reader[0][i].todense())
 # not ready #                 for j in range(neta):
